Switched libraries and fixed forced mode readings
Changed to the manchester encoding library since the other ones wouldn't work at 1Mhz and were also causing the delay() function to stop working on the AtTiny. Am still able to send from one side of the lab to the other, though tomorrow I plan on installing an antenna. Forced mode readings were also fixed--the sensor must be placed into forced mode each time that you want the reading to be updated (since it is stored), which I didn't realize until I read the datasheet more thoroughly. Humidity readings need to be fixed as they are very inaccurate and are fluctuating wildly.