Create AtMega_RFID_Tag_Test
Code to test the ST25DV04K. The existing library for this chip requires several other libraries to work and is too complicated for what I'm trying to do, so I'm making my own functions instead. Presenting the I2C password successfully unlocks protected mode (according to the register) and I've verified that the program is able to read some various normal registers.