Changed to transmit only with two transmitting modes
When the switch is off, the sensor will transmit in brief mode, where the sensor will send out an "okay" pulse and its address (currently it sends this out 5 times). When the switch is on, the sensor sends out its readings like normal. The sensor was able to send a good distance from within a plastic shell with a power supply of 3 AA batteries.